\name{randomSeed}
\alias{randomSeed}



\title{
Random seed generator
}


\description{
This function generates a random seed based on the system time.
}


\usage{
randomSeed()
}



\value{
 \item{value}{an integer value}
}


\author{
Chuang Ma, Xiangfeng Wang
}



\examples{
   randomSeed()
}



\keyword{ statistic }

